backend/pyproject.toml
[tool.poetry]
name = "stocker"
version = "0.1.0"
description = "Systematic Trading Platform - Volatility-Targeted Trend-Following Strategy"
authors = ["Daniel Jackson"]
readme = "README.md"
packages = [{include = "stocker"}]

[tool.poetry.dependencies]
python = "^3.12"
# Core Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"

# Database
sqlalchemy = "^2.0.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.0"

# Redis & Async
redis = {extras = ["asyncio"], version = "^5.2.0"}
celery = "^5.4.0"

# Data Processing
pandas = "^2.2.0"
numpy = "^2.1.0"

# HTTP Client
httpx = "^0.27.0"

# Configuration
python-dotenv = "^1.0.0"

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
greenlet = "^3.3.0"
alpaca-py = "^0.43.2"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"  # For testing FastAPI

# Linting & Formatting
ruff = "^0.8.0"
mypy = "^1.13.0"

# Type Stubs
types-redis = "^4.6.0"
types-passlib = "^1.7.0"
pandas-stubs = "^2.2.0"

[tool.poetry.group.backtest.dependencies]
# Backtesting & Analysis
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
yfinance = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "yfinance.*",
    "seaborn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=stocker",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["stocker"]
omit = [
    "*/tests/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
